Renamed ParsedDocumentBuilder for Asn1SccParsedDocumentBuilder and ParsedDocumentBuilderInterface for ParsedDocumentBuilder
