+++ asn1sccparseddocumentbuilder.cpp
@@ -23,7 +23,7 @@
 **
 ****************************************************************************/
 
-#include "parseddocumentbuilder.h"
+#include "asn1sccparseddocumentbuilder.h"
 
 #include <QNetworkReply>
 #include <QJsonDocument>
@@ -42,27 +42,27 @@
 
 using namespace Asn1Acn::Internal;
 
-ParsedDocumentBuilderInterface *ParsedDocumentBuilder::create(const QHash<QString, DocumentSourceInfo> &documents)
+ParsedDocumentBuilder *Asn1SccParsedDocumentBuilder::create(const QHash<QString, DocumentSourceInfo> &documents)
 {
     auto serviceProvider = ExtensionSystem::PluginManager::getObject<ParsingServiceProvider>();
-    return new ParsedDocumentBuilder(documents, serviceProvider);
+    return new Asn1SccParsedDocumentBuilder(documents, serviceProvider);
 }
 
-ParsedDocumentBuilder::ParsedDocumentBuilder(const QHash<QString, DocumentSourceInfo> &documents,
-                                             ParsingServiceProvider *serviceProvider)
+Asn1SccParsedDocumentBuilder::Asn1SccParsedDocumentBuilder(const QHash<QString, DocumentSourceInfo> &documents,
+                                                           ParsingServiceProvider *serviceProvider)
     : m_serviceProvider(serviceProvider), m_rawDocuments(documents)
 {
 }
 
-void ParsedDocumentBuilder::run()
+void Asn1SccParsedDocumentBuilder::run()
 {
     QNetworkReply *reply = m_serviceProvider->requestAst(m_rawDocuments);
 
     QObject::connect(reply, &QNetworkReply::finished,
-                     this, &ParsedDocumentBuilder::requestFinished);
+                     this, &Asn1SccParsedDocumentBuilder::requestFinished);
 }
 
-void ParsedDocumentBuilder::requestFinished()
+void Asn1SccParsedDocumentBuilder::requestFinished()
 {
     QNetworkReply *reply = qobject_cast<QNetworkReply *>(sender());
 
@@ -74,7 +74,7 @@ void ParsedDocumentBuilder::requestFinished()
     reply->deleteLater();
 }
 
-void ParsedDocumentBuilder::parseResponse(const QByteArray &jsonData)
+void Asn1SccParsedDocumentBuilder::parseResponse(const QByteArray &jsonData)
 {
     const auto json = QJsonDocument::fromJson(jsonData).object();
     if (responseContainsAst(json)) {
@@ -87,7 +87,7 @@ void ParsedDocumentBuilder::parseResponse(const QByteArray &jsonData)
     }
 }
 
-void ParsedDocumentBuilder::parseXML(const QString &textData)
+void Asn1SccParsedDocumentBuilder::parseXML(const QString &textData)
 {
     QXmlStreamReader reader;
     reader.addData(textData);
@@ -108,23 +108,23 @@ void ParsedDocumentBuilder::parseXML(const QString &textData)
     }
 }
 
-std::vector<std::unique_ptr<ParsedDocument>> ParsedDocumentBuilder::takeDocuments()
+std::vector<std::unique_ptr<ParsedDocument>> Asn1SccParsedDocumentBuilder::takeDocuments()
 {
     return std::move(m_parsedDocuments);
 }
 
-bool ParsedDocumentBuilder::responseContainsAst(const QJsonObject &json)
+bool Asn1SccParsedDocumentBuilder::responseContainsAst(const QJsonObject &json)
 {
     return json[QLatin1Literal("ErrorCode")].toInt(-1) == 0;
 }
 
-void ParsedDocumentBuilder::storeErrorMessages(const QJsonObject &json)
+void Asn1SccParsedDocumentBuilder::storeErrorMessages(const QJsonObject &json)
 {
     for (const auto message : json[QLatin1Literal("Messages")].toArray())
         m_errorMessages.append(message.toString());
 }
 
-QString ParsedDocumentBuilder::getAstXml(const QJsonObject &json)
+QString Asn1SccParsedDocumentBuilder::getAstXml(const QJsonObject &json)
 {
     const auto files = json[QLatin1Literal("Files")].toArray();
     if (files.size() != 1)
