+++ documentprocessor.cpp
@@ -28,7 +28,7 @@
 #include <QFileInfo>
 
 #include "parseddocument.h"
-#include "parseddocumentbuilder.h"
+#include "asn1sccparseddocumentbuilder.h"
 
 using namespace Asn1Acn::Internal;
 
@@ -54,7 +54,7 @@ void DocumentProcessor::addToRun(const QString &docContent, const QString &fileP
 
 void DocumentProcessor::run()
 {
-    m_docBuilder = ParsedDocumentBuilder::create(m_documents);
+    m_docBuilder = Asn1SccParsedDocumentBuilder::create(m_documents);
     m_docBuilder->run();
 
     QObject::connect(dynamic_cast<QObject *>(m_docBuilder), SIGNAL(finished()), this, SLOT(onBuilderFinished()));
